datatype correction

+++ config/clickhouse/scripts/drug_warnings.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS drug_warnings ENGINE = EmbeddedRocksDB () PRIMARY KEY
             chemblIds Array (String),
             country Nullable (String),
             description Nullable (String),
-            id Nullable (UInt64),
+            id Nullable (UInt32),
             references Array (
                 Tuple (
                     ref_id String,
